Add files via upload
In this exercise, you will be writing a program to calculate the total cost of a meal with tax and tip. You need to do the following:
Create variables to hold the sales tax rate (8.25%).
Create a variable to hold the percent you want to tip (20% for example).
Create a variable whose value is a user input for the cost before tax and tip.
Based on the user input, calculate the sales tax, tip, and add them to the user input cost.
Print your final result.